core: Move mpi_iccp3m_init()
